refactor: rename parser directory
